Abstract root checking
